app/requirements.txt
av==10.0.0
efficientnet-pytorch==0.7.1
facenet-pytorch==2.5.3
fastapi==0.98.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
mpmath==1.3.0
networkx==3.1
Pillow==9.5.0
pydantic==1.10.9
starlette==0.27.0
python-multipart==0.0.6
uvicorn==0.22.0
pandas==2.0.3
streamlit==1.24.0
